shopify_returns_chat_agent/tests/integration/requirements.txt
# Integration Testing Requirements for Shopify Returns Chat Agent

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-json-report>=1.5.0
pytest-html>=3.2.0
pytest-cov>=4.1.0

# HTTP client libraries
requests>=2.31.0
aiohttp>=3.8.5
httpx>=0.24.1

# Environment and configuration
python-dotenv>=1.0.0

# Data handling and utilities
pydantic>=2.1.1
uuid>=1.30

# Performance and monitoring
psutil>=5.9.5
memory-profiler>=0.60.0

# Concurrent processing
concurrent-futures>=3.1.1 # For Python < 3.2 compatibility

# Statistics and data analysis
statistics>=1.0.3.5 # Built-in but ensuring compatibility

# Optional: Database testing (if needed)
# sqlalchemy>=2.0.0
# pytest-postgresql>=5.0.0

# Optional: Mock external services  
responses>=0.23.0
pytest-mock>=3.11.0

# Optional: Load testing utilities
locust>=2.16.0 